+++ include/tantivy_search_cxx.h
@@ -958,17 +958,17 @@ struct Statistics final {
 // - `topk`: only return top k related results.
 // - `alived_ids`: alived rowIds given by u32 vector.
 // - `query_with_filter`: whether use alived_bitmap or not.
-::BM25Result ffi_bm25_search_filter_ids(::std::string const &index_path, ::std::string const &sentence, ::std::uint32_t topk, ::std::vector<::std::uint32_t> const &alived_ids, bool query_with_filter) noexcept;
+::BM25Result ffi_bm25_search_filter_ids(::std::string const &index_path, ::std::string const &sentence, ::std::uint32_t topk, ::std::vector<::std::uint64_t> const &alived_ids, bool query_with_filter) noexcept;
 
 // Execute a regex query and return rowIds u8 bitmap.
 // arguments:
 // - `index_path`: index directory.
-// - `sentence`: from ClickHouse TextSearch function.
+// - `sentence`: from DingoDB TextSearch function. If search for bytes field, please use base64 encode for bytes.
 // - `topk`: only return top k related results.
 // - `alived_ids`: alived rowIds given by u32 vector.
 // - `query_with_filter`: whether use alived_bitmap or not.
 // - `colunm_names`: for multi column search.
-::BM25Result ffi_bm25_search_with_column_names(::std::string const &index_path, ::std::string const &sentence, ::std::uint32_t topk, ::std::vector<::std::uint32_t> const &alived_ids, bool query_with_filter, ::std::vector<::std::string> const &column_names) noexcept;
+::BM25Result ffi_bm25_search_with_column_names(::std::string const &index_path, ::std::string const &sentence, ::std::uint32_t topk, ::std::vector<::std::uint64_t> const &alived_ids, bool query_with_filter, ::std::vector<::std::string> const &column_names) noexcept;
 
 // Get doc freq for current part.
 // arguments:
